+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Pass `allow-downgrade` flag to `rustup` if `nightly` toolchain with components requested
+
 ## [1.0.5] - 2020-01-26
 
 ### Fixed